metrics: Expose endpoint and policy computation time metrics
This commit adds changes to export endpoint and policy computation time
metrics to prometheus. To have an idea of the distribution, assuming a normal form, we need all 3:

1. count (policy or endpoint)
2. sum of time taken to regenerate
3. sum of squares of time taken to regenerate

Policy and Endpoint count are already exposed as follows:
1. EndpointRegenerationCount gives endpoint regenerations that have completed.
2. PolicyCount and PolicyImportErrors gives total successful policies imported.

This commit adds sum of time taken to regenerate and sum of squares of time taken to regenerate
for policy generation/computation and endpoint regeneration.

Fixes: #4396
Signed-Off-By: Manali Bhutiyani <manali@covalent.io>